Add files via upload
1.)Project Introduction
This project, called Skychain, aims to combine blockchain technology and cloud
file storage competence. Combined with the use of interplanetary file storage, abbreviated
IPFS, which provides superior security thanks to the transactions that blockchain connects
to via metamask and executes the upload and allow functions while executing the upload
and allow functions, these website files are both decentralized and fragmented. It brings a
new breath to storage.
This project, prepared using Metamask, Ether.js, Solidity, React.js, Hardhat, IPFS
and Pinata API, makes it very easy to use the website it connects to with the technologies it
uses. This project, for which I wrote the smart contract with the Ethereum-based Solidity
language, can upload files to the system, view them in the system, and download them
from the system. It includes capabilities such as deleting from the system and allowing
other accounts.
We run this project with three terminals that we use in visual studio code. npx
hardhat node command allows us to run our smart contract and then view the activities
we do in this terminal. The npx hardhat run scripts/deploy.js –network localhost
command deploys this smart contract to a certain contract address so that we can use it on
our website. And finally, since we are using react.js, we need to open our website in the
terminal. For this, click npm start after the cd client. We have opened our website on
localhost.